# course/OpenAPI definition.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "description": "",
    "version": "v0"
  },
  "tags": [
    {
      "name": "course-controller"
    },
    {
      "name": "student-controller"
    },
    {
      "name": "enrollment-controller"
    }
  ],
  "paths": {
    "/api/courses/{id}": {
      "get": {
        "summary": "getCourseById",
        "deprecated": false,
        "description": "",
        "operationId": "getCourseById",
        "tags": [
          "course-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateCourse",
        "deprecated": false,
        "description": "",
        "operationId": "updateCourse",
        "tags": [
          "course-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteCourse",
        "deprecated": false,
        "description": "",
        "operationId": "deleteCourse",
        "tags": [
          "course-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/courses": {
      "get": {
        "summary": "getAllCourses",
        "deprecated": false,
        "description": "",
        "operationId": "getAllCourses",
        "tags": [
          "course-controller"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "createCourse",
        "deprecated": false,
        "description": "",
        "operationId": "createCourse",
        "tags": [
          "course-controller"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/students/{id}": {
      "get": {
        "summary": "getStudentById",
        "deprecated": false,
        "description": "",
        "operationId": "getStudentById",
        "tags": [
          "student-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateStudent",
        "deprecated": false,
        "description": "",
        "operationId": "updateStudent",
        "tags": [
          "student-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteStudent",
        "deprecated": false,
        "description": "",
        "operationId": "deleteStudent",
        "tags": [
          "student-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/students": {
      "get": {
        "summary": "getAllStudents",
        "deprecated": false,
        "description": "",
        "operationId": "getAllStudents",
        "tags": [
          "student-controller"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "createStudent",
        "deprecated": false,
        "description": "",
        "operationId": "createStudent",
        "tags": [
          "student-controller"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/enrollments": {
      "get": {
        "summary": "getAllEnrollments",
        "deprecated": false,
        "description": "",
        "operationId": "getAllEnrollments",
        "tags": [
          "enrollment-controller"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "enrollStudent",
        "deprecated": false,
        "description": "",
        "operationId": "enrollStudent",
        "tags": [
          "enrollment-controller"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/enrollments/student/{studentId}": {
      "get": {
        "summary": "getByStudent",
        "deprecated": false,
        "description": "",
        "operationId": "getByStudent",
        "tags": [
          "enrollment-controller"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/enrollments/course/{courseId}": {
      "get": {
        "summary": "getByCourse",
        "deprecated": false,
        "description": "",
        "operationId": "getByCourse",
        "tags": [
          "enrollment-controller"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/enrollments/{id}": {
      "delete": {
        "summary": "deleteEnrollment",
        "deprecated": false,
        "description": "",
        "operationId": "deleteEnrollment",
        "tags": [
          "enrollment-controller"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "Course": {
        "required": [
          "code",
          "instructor",
          "schedule",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "instructor": {
            "$ref": "#/components/schemas/Instructor"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleSlot"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "enrolled": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Instructor": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ScheduleSlot": {
        "required": [
          "dayOfWeek",
          "endTime",
          "startTime"
        ],
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "expectedAttendance": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Student": {
        "required": [
          "email",
          "grade",
          "major",
          "name",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "major": {
            "type": "string"
          },
          "grade": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}